int32_t->int64_t

+++ oneflow/core/kernel/accuracy_print_kernel.cpp
@@ -10,7 +10,7 @@ void AccuracyPrintKernel<T>::Forward(const KernelCtx& kernel_ctx,
   const Blob* accuracy_acc_blob = BnInOp2Blob("accuracy_acc");
   const Blob* accuracy_instance_num_blob = BnInOp2Blob("accuracy_instance_num");
   T accurate_num = accuracy_acc_blob->dptr<T>()[0];
-  int32_t accuracy_instance_num = static_cast<int32_t>(accuracy_instance_num_blob->dptr<T>()[0]);
+  int64_t accuracy_instance_num = static_cast<int64_t>(accuracy_instance_num_blob->dptr<T>()[0]);
   float accuracy = accurate_num / accuracy_instance_num;
   const char* accuracy_op_name = op_conf().name().c_str() + AccuracyPrintPrefix.length();
   auto kernel_conf = this->kernel_conf();
