+++ oneflow/core/kernel/kernel.cpp
@@ -197,7 +197,7 @@ template<DeviceType device_type, typename T>
 void KernelIfWithModel<device_type, T>::SetTotalInstanceNumDiffBlob(
     const KernelCtx& ctx, const std::function<Blob*(const std::string&)>& BnInOp2Blob) const {
   CHECK_GE(this->op_attribute().model_bns().size(), 2);
-  int32_t dim0_valid_num_sum =
+  int64_t dim0_valid_num_sum =
       BnInOp2Blob(this->op_attribute().output_diff_bns(0))->CalcDim0ValidNumSum();
   Blob* total_instance_num_diff_blob = BnInOp2Blob("total_instance_num_diff");
   KernelUtil<device_type, T>::Set(ctx.device_ctx, static_cast<T>(dim0_valid_num_sum),
