+++ oneflow/core/kernel/loss_kernel.cpp
@@ -6,7 +6,7 @@ template<DeviceType device_type, typename PredType, typename LabelType>
 void LossKernel<device_type, PredType, LabelType>::SetLossInstanceNumBlob(
     const KernelCtx& ctx, const std::function<Blob*(const std::string&)>& BnInOp2Blob) const {
   CHECK_GE(this->op_attribute().input_bns().size(), 2);
-  int32_t dim0_valid_num_sum =
+  int64_t dim0_valid_num_sum =
       BnInOp2Blob(this->op_attribute().input_bns(0))->CalcDim0ValidNumSum();
   KernelUtil<device_type, PredType>::Set(ctx.device_ctx, static_cast<PredType>(dim0_valid_num_sum),
                                          BnInOp2Blob("loss_instance_num")->mut_dptr<PredType>());
